Refactor according to typings structure
